Refactor py script for circom main files and inputs
